Added Text API on server and client now loads map from local files

+++ server/api/src/main/java/me/eddiep/ghost/common/network/packet/SpawnEntityPacket.java
@@ -23,7 +23,7 @@ public class SpawnEntityPacket extends Packet<BaseServer, BasePlayerClient> {
         write((byte)0x10)
                 .write(type)
                 .write(toSpawn.getID())
-                .write((byte)toSpawn.getName().length())
+                .write(toSpawn.getName().length())
                 .write(toSpawn.getName())
                 .write(toSpawn.getX())
                 .write(toSpawn.getY())
